components/com.iot.WaterPump/requirements.txt
paho-mqtt
python-dotenv
python-decouple
RPi.GPIO
